@@ -8,16 +8,17 @@
   "license": "MIT",
   "keywords": [],
   "devDependencies": {
-    "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.10",
-    "babel-preset-react": "^6.22.0",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-react": "^6.24.1",
     "json-loader": "^0.5.4",
-    "react": "^15.4.2",
+    "react": "^15.5.4",
     "react-hot-loader": "^1.3.1",
-    "webpack": "^2.2.0",
-    "webpack-dev-server": "^1.16.2"
+    "webpack": "^2.5.1",
+    "webpack-dev-server": "^2.4.5"
   },
   "dependencies": {
-    "browserify-zlib-next": "^1.0.1"
+    "browserify-zlib-next": "^1.0.1",
+    "safe-buffer": "^5.0.1"
   }
 }
